refactor: clean up tool structure
