+++ package.json
@@ -56,6 +56,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5.8.2"
+    "typescript": "^5.8.2",
+    "zod-to-json-schema": "^3.24.4"
   }
 }
