+++ src/tools/morpho/handlers.ts
@@ -1,17 +1,16 @@
 import type { WalletClient } from 'viem';
 import type { z } from 'zod';
-import type { GetMorphoVaultsSchema } from '../schemas.js';
-import { getMorphoVaults } from './index.js';
-import type { MorphoVault } from './types.js';
+import { GetMorphoVaultsSchema } from './schemas.js';
+import { getMorphoVaults } from './utils.js';
 
 export async function getMorphoVaultsHandler(
   wallet: WalletClient,
   args: z.infer<typeof GetMorphoVaultsSchema>,
-): Promise<MorphoVault[]> {
+): Promise<string> {
   const vaults = await getMorphoVaults({
     chainId: wallet.chain?.id ?? 8453,
     assetSymbol: args.assetSymbol ?? '',
   });
 
-  return vaults;
+  return JSON.stringify(vaults);
 }
