+++ src/tools/morpho/index.ts
@@ -1,22 +1,10 @@
-import { base } from 'viem/chains';
-import { morphoApiClient } from './graphql.js';
-import { GET_VAULTS_QUERY } from './queries.js';
-import type { MorphoVaultsResponse } from './types.js';
+import { generateTool } from '../../utils.js';
+import { getMorphoVaultsHandler } from './handlers.js';
+import { GetMorphoVaultsSchema } from './schemas.js';
 
-export async function getMorphoVaults({
-  chainId = base.id,
-  assetSymbol,
-}: {
-  chainId?: number;
-  assetSymbol: string;
-}) {
-  const data = await morphoApiClient.request<MorphoVaultsResponse>(
-    GET_VAULTS_QUERY,
-    {
-      chainId,
-      assetSymbol,
-    },
-  );
-
-  return data.vaults.items;
-}
+export const getMorphoVaultsTool = generateTool({
+  name: 'get_morpho_vaults',
+  description: 'Get the vaults available for a particular asset on Morpho',
+  inputSchema: GetMorphoVaultsSchema,
+  toolHandler: getMorphoVaultsHandler,
+});
