+++ src/tools/types.ts
@@ -1,44 +1,9 @@
 /**
- * Patches for Onramp
- */
-
-type OnrampPaymentMethod = {
-  id: string;
-};
-
-type OnrampConfigResponseData = {
-  countries: OnrampConfigCountry[];
-};
-
-type OnrampConfigCountry = {
-  id: string;
-  subdivisions: string[];
-  paymentMethods: OnrampPaymentMethod[];
-};
-
-type TokenNetwork = {
-  name: string;
-  display_name: string;
-  contract_address: string;
-  chain_id: string;
-  icon_url: string;
-};
-
-type TokenInfo = {
-  id: string;
-  name: string;
-  symbol: string;
-  networks: TokenNetwork[];
-  icon_url: string;
-};
-
-export type PatchedOnrampConfigResponseData = OnrampConfigResponseData & {
-  purchase_currencies: TokenInfo[];
-};
-
-/**
  * OpenRouter types
  */
+
+import type { Tool } from '@modelcontextprotocol/sdk/types.js';
+import type { PublicActions, WalletClient } from 'viem';
 
 export type OpenRouterTransferIntentResponse = {
   data: {
@@ -67,4 +32,14 @@ export type OpenRouterTransferIntentResponse = {
       };
     };
   };
+};
+
+export type ToolHandler = (
+  wallet: WalletClient & PublicActions,
+  args: any,
+) => Promise<string>;
+
+export type ToolWithHandler = {
+  definition: Tool;
+  handler: ToolHandler;
 };
